TokencodeWidgetService: Use REORDER_TO_FRONT to fix activity stack
Repeatedly clicking <widget> <home> would create multiple copies of
MainActivity on the stack, and not all of them would necessarily be in
sync (e.g. with the current PIN).  Use FLAG_ACTIVITY_REORDER_TO_FRONT
to ensure that there is only one copy of MainActivity when started from
the PendingIntent.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>