TokencodeWidgetService: Make our PendingIntent work like a launcher click
When the user clicks on the tokencode widget, it (re)starts MainActivity.
Under some circumstances this can result in a task history that involves
duplicate copies of the activity, e.g.

  Stack #1:
    Task id #1088
      TaskRecord{426608e8 #1088 A=app.easytoken U=0 sz=2}
      Intent { flg=0x10000000 cmp=app.easytoken/.MainActivity }
        Hist #1: ActivityRecord{419e21d0 u0 app.easytoken/.MainActivity t1088}
          Intent { act=android.intent.action.MAIN cat=[android.intent.category.LAUNCHER] flg=0x10600000 cmp=app.easytoken/.MainActivity bnds=[540,853][804,1152] }
          ProcessRecord{4290c760 10102:app.easytoken/u0a86}
        Hist #0: ActivityRecord{42660700 u0 app.easytoken/.MainActivity t1088}
          Intent { flg=0x10000000 cmp=app.easytoken/.MainActivity bnds=[86,919][465,1085] }

If we start MainActivity with the same intent settings as the Launcher,
it will operate as though the user clicked the Launcher icon.  This is
exactly what we want.

Reference:

http://stackoverflow.com/questions/5502427/resume-application-and-stack-from-notification

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>