ImportActivity: Always confirm token import
Importing from an email client can be somewhat disconcerting, as tokens
with no devid/password will be silently accepted with no user interaction.
Then if the user clicks on the token again, he will be prompted to
overwrite the "old" token (which he just imported) with the same "new"
token.

So, let's display ImportConfirmFragment unconditionally.  For new tokens
the verbiage will be slightly different, but at least it will provide
a clear way to accept/decline the new token.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>